+++ src/cpp/session/modules/SessionHelp.R
@@ -122,11 +122,29 @@ options(help_type = "html")
       error = function(e) NULL
    )
    
-   objectNames <- c(
-      objects(ns, all.names = TRUE),
-      unname(grep(" ", datasets$results[, "Item"], fixed = TRUE, value = TRUE, invert = TRUE))
+   objectNames <- objects(ns, all.names = TRUE)
+   datasetNames <- unname(grep(" ", datasets$results[, "Item"], fixed = TRUE, value = TRUE, invert = TRUE))
+   
+   objects <- tryCatch(
+      mget(objectNames, envir = ns, inherits = TRUE),
+      error = function(e) NULL
    )
-   objects <- mget(objectNames, envir = ns, inherits = TRUE)
+   
+   # Try to get the datasets from the namespace. These will only exist if they have
+   # been explicitly loaded, so be careful to tryCatch here.
+   data <- lapply(datasetNames, function(x) {
+      tryCatch(
+         get(x, envir = ns),
+         error = function(e) NULL
+      )
+   })
+   
+   # Combine them together
+   if (length(data))
+   {
+      objects <- c(objects, data)
+      objectNames <- c(objectNames, datasetNames)
+   }
    
    # Find which object is actually identical to the one we have
    success <- FALSE
@@ -186,6 +204,12 @@ options(help_type = "html")
    if (!length(type))
       return()
    
+   # Avoid install.packages hook
+   if (what == "install.packages" &&
+       type == .rs.acCompletionTypes$ARGUMENT &&
+       is.null(from))
+      return(.rs.getHelp("install.packages", "utils"))
+   
    # Help for options
    if (type == .rs.acCompletionTypes$OPTION)
       return(.rs.getHelp("options", "base", subset = FALSE))
@@ -225,7 +249,7 @@ options(help_type = "html")
    # If 'src' is the name of something on the searchpath, get that object
    # from the seach path, then attempt to get help based on that object
    pos <- match(src, search(), nomatch = -1L)
-   if (pos > 0)
+   if (pos >= 0)
    {
       object <- tryCatch(get(name, pos = pos), error = function(e) NULL)
       if (!is.null(object))
@@ -266,7 +290,7 @@ options(help_type = "html")
    {
       pos <- match(src, search(), nomatch = -1L)
       
-      if (pos > 0)
+      if (pos >= 0)
       {
          object <- tryCatch(get(functionName, pos = pos), error = function(e) NULL)
          if (!is.null(object))
@@ -311,13 +335,16 @@ options(help_type = "html")
    
    # If 'package' is the name of something on the search path, then we
    # attempt to resolve the object and get its help.
-   pos <- match(package, search(), nomatch = -1L)
-   if (pos > 0)
+   if (length(package))
    {
-      object <- tryCatch(get(topic, pos = pos), error = function(e) NULL)
-      if (is.null(object))
-         return(NULL)
-      return(.rs.getHelpFromObject(object, envir))
+      pos <- match(package, search(), nomatch = -1L)
+      if (pos >= 0)
+      {
+         object <- tryCatch(get(topic, pos = pos), error = function(e) NULL)
+         if (is.null(object))
+            return(NULL)
+         return(.rs.getHelpFromObject(object, envir))
+      }
    }
    
    helpfiles <- NULL