+++ src/gwt/src/org/rstudio/studio/client/RStudioGinjector.java
@@ -66,6 +66,7 @@ import org.rstudio.studio.client.workbench.views.source.editors.text.TextEditing
 import org.rstudio.studio.client.workbench.views.source.editors.text.TextEditingTargetRMarkdownHelper;
 import org.rstudio.studio.client.workbench.views.source.editors.text.cpp.CppCompletionManager;
 import org.rstudio.studio.client.workbench.views.source.editors.text.cpp.CppCompletionRequest;
+import org.rstudio.studio.client.workbench.views.source.editors.text.r.RCompletionToolTip;
 import org.rstudio.studio.client.workbench.views.vcs.svn.SVNCommandHandler;
 import org.rstudio.studio.client.workbench.views.environment.ClearAllDialog;
 
@@ -100,6 +101,7 @@ public interface RStudioGinjector extends Ginjector
    void injectMembers(LocalRepositoriesWidget localRepositoriesWidget);
    void injectMembers(CppCompletionRequest request);
    void injectMembers(CppCompletionManager completionManager);
+   void injectMembers(RCompletionToolTip toolTip);
    
    public static final RStudioGinjector INSTANCE = GWT.create(RStudioGinjector.class);