+++ src/gwt/src/org/rstudio/studio/client/application/ui/GlobalToolbar.java
@@ -14,6 +14,7 @@
  */
 package org.rstudio.studio.client.application.ui;
 
+import org.rstudio.core.client.StringUtil;
 import org.rstudio.core.client.theme.res.ThemeResources;
 import org.rstudio.core.client.widget.CanFocus;
 import org.rstudio.core.client.widget.FocusContext;
@@ -176,6 +177,16 @@ public class GlobalToolbar extends Toolbar
       ProjectPopupMenu projectMenu = new ProjectPopupMenu(sessionInfo,
                                                           commands_);
       addRightWidget(projectMenu.getToolbarButton());
+      
+      // go to project dir button
+      if (!StringUtil.isNullOrEmpty(
+            sessionInfo.getActiveProjectFile()))
+      {
+         addRightSeparator();
+         addRightWidget(commands_.setWorkingDirToProjectDir().createToolbarButton());
+      }
+      
+      
    }
    
    @Override