+++ src/gwt/src/org/rstudio/studio/client/workbench/views/console/shell/assist/CompletionPopupPanel.java
@@ -22,9 +22,8 @@ import com.google.gwt.event.logical.shared.CloseHandler;
 import com.google.gwt.event.logical.shared.SelectionEvent;
 import com.google.gwt.event.logical.shared.SelectionHandler;
 import com.google.gwt.event.shared.HandlerRegistration;
-import com.google.gwt.user.client.Event;
+import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.HTML;
-import com.google.gwt.user.client.ui.HorizontalPanel;
 import com.google.gwt.user.client.ui.PopupPanel;
 
 import org.rstudio.core.client.ElementIds;
@@ -41,22 +40,31 @@ public class CompletionPopupPanel extends ThemedPopupPanel
 {
    public CompletionPopupPanel()
    {
-      super() ;
+      super();
       styles_ = ConsoleResources.INSTANCE.consoleStyles();
-      help_ = new HelpInfoPane();
+      help_ = new HelpInfoPopupPanel();
       help_.setWidth("400px");
-      help_.clearHelp(false);
-      help_.setVisible(false);
+      help_.show();
+      
       setStylePrimaryName(styles_.completionPopup()) ;
+      
+      hideAll();
+      
       addCloseHandler(new CloseHandler<PopupPanel>() {
          
          @Override
          public void onClose(CloseEvent<PopupPanel> event)
          {
-            help_.clearHelp(false);
-            help_.setVisible(false);
+            hideAll();
          }
       });
+   }
+   
+   private void hideAll()
+   {
+      // Throw everything off-screen to reduce flickering
+      setVisible(false);
+      help_.setVisible(false);
    }
 
    public void showProgress(String progress, PositionCallback callback)
@@ -83,33 +91,32 @@ public class CompletionPopupPanel extends ThemedPopupPanel
    {
       CompletionList<QualifiedName> list = new CompletionList<QualifiedName>(
                                        values,
-                                       7,
+                                       6,
                                        true,
-                                       false) ;
+                                       true) ;
 
       list.addSelectionCommitHandler(new SelectionCommitHandler<QualifiedName>() {
          public void onSelectionCommit(SelectionCommitEvent<QualifiedName> event)
          {
+            lastSelectedValue_ = event.getSelectedItem();
             SelectionCommitEvent.fire(CompletionPopupPanel.this, 
                                       event.getSelectedItem()) ;
          }
-      }) ;
+      });
+      
       list.addSelectionHandler(new SelectionHandler<QualifiedName>() {
          public void onSelection(SelectionEvent<QualifiedName> event)
          {
+            lastSelectedValue_ = event.getSelectedItem();
             SelectionEvent.fire(CompletionPopupPanel.this, 
                                 event.getSelectedItem()) ;
          }
-      }) ;
+      });
+      
       list_ = list ;
+      setWidget(list_);
       
-      HorizontalPanelWithMouseEvents horiz 
-                                 = new HorizontalPanelWithMouseEvents() ;
-      horiz.add(list_) ;
-      horiz.add(help_) ;
-      
-      setWidget(horiz) ;
-      ElementIds.assignElementId(horiz.getElement(), 
+      ElementIds.assignElementId(list_.getElement(), 
             ElementIds.POPUP_COMPLETIONS);
       
       show(callback) ;
@@ -127,6 +134,8 @@ public class CompletionPopupPanel extends ThemedPopupPanel
          setPopupPositionAndShow(callback) ;
       else
          show() ;
+      if (help_ != null)
+         resolveHelpPosition(help_.isVisible());
    }
    
    public QualifiedName getSelectedValue()
@@ -135,6 +144,11 @@ public class CompletionPopupPanel extends ThemedPopupPanel
          return null ;
       
       return list_.getSelectedItem() ;
+   }
+   
+   public QualifiedName getLastSelectedValue()
+   {
+      return lastSelectedValue_;
    }
    
    public Rectangle getSelectionRect()
@@ -180,25 +194,58 @@ public class CompletionPopupPanel extends ThemedPopupPanel
    @Override
    public void displayHelp(ParsedInfo help)
    {
-      help_.setVisible(help.hasInfo());
+      if (list_ == null || !list_.isAttached())
+         return;
+      
       help_.displayHelp(help) ;
-      help_.setHeight(list_.getOffsetHeight() + "px") ;
+      resolveHelpPosition(help.hasInfo());
    }
    
    @Override
    public void displayParameterHelp(Map<String, String> map, String parameterName)
    {
-      help_.setVisible(map.get(parameterName) != null);
+      if (list_ == null || !list_.isAttached())
+         return;
+      
       help_.displayParameterHelp(map, parameterName) ;
-      help_.setHeight(list_.getOffsetHeight() + "px") ;
+      resolveHelpPosition(map.get(parameterName) != null);
    }
    
    @Override
    public void displayPackageHelp(ParsedInfo help)
    {
-      help_.setVisible(help.hasInfo());
-      help_.displayPackageHelp(help) ;
-      help_.setHeight(list_.getOffsetHeight() + "px") ;
+      if (list_ == null || !list_.isAttached())
+         return;
+      
+      help_.displayPackageHelp(help);
+      resolveHelpPosition(help.hasInfo());
+      
+   }
+   
+   private void resolveHelpPosition(boolean setVisible)
+   {
+      int top = getAbsoluteTop();
+      int left = getAbsoluteLeft();
+      int bottom = top + getOffsetHeight();
+      int width = getOffsetWidth();
+      
+      // If displaying the help with the top aligned to the completion list
+      // would place the help offscreen, then re-align it so that the bottom of the
+      // help is aligned with the completion popup.
+      
+      if (!help_.isShowing())
+         help_.show();
+      
+      // NOTE: Help has not been positioned yet so what we're really asking is,
+      // 'if we align the top of help with the top of the completion list, will
+      // it flow offscreen?'
+      
+      if (top + help_.getOffsetHeight() + 20 > Window.getClientHeight())
+         top = bottom - help_.getOffsetHeight()
+               - 9; // fudge factor
+      
+      help_.setPopupPosition(left + width - 4, top + 3);
+      help_.setVisible(setVisible);
    }
    
    @Override
@@ -228,7 +275,12 @@ public class CompletionPopupPanel extends ThemedPopupPanel
    {
       return addDomHandler(handler, MouseDownEvent.getType()) ;
    }
-
+   
+   public boolean isHelpVisible()
+   {
+      return help_.isVisible() && help_.isShowing();
+   }
+   
    private HTML setText(String text)
    {
       HTML contents = new HTML() ;
@@ -237,39 +289,8 @@ public class CompletionPopupPanel extends ThemedPopupPanel
       return contents ;
    }
    
-   private static class HorizontalPanelWithMouseEvents 
-         extends HorizontalPanel 
-         implements HasMouseOverHandlers, 
-                    HasMouseOutHandlers
-   {
-      public HorizontalPanelWithMouseEvents()
-      {
-         super() ;
-         sinkEvents(Event.ONMOUSEOVER 
-                  | Event.ONMOUSEOUT 
-                  | Event.ONMOUSEDOWN
-                  | Event.ONFOCUS
-                  | Event.ONBLUR) ;
-      }
-
-      public HandlerRegistration addMouseOverHandler(MouseOverHandler handler)
-      {
-         return addHandler(handler, MouseOverEvent.getType()) ;
-      }
-
-      public HandlerRegistration addMouseOutHandler(MouseOutHandler handler)
-      {
-         return addHandler(handler, MouseOutEvent.getType()) ;
-      }
-      
-      @SuppressWarnings("unused")
-      public HandlerRegistration addMouseDownHandler(MouseDownHandler handler)
-      {
-         return addHandler(handler, MouseDownEvent.getType()) ;
-      }
-   }
-   
    private CompletionList<QualifiedName> list_ ;
-   private HelpInfoPane help_ ;
+   private HelpInfoPopupPanel help_ ;
    private final ConsoleResources.ConsoleStyles styles_;
+   private static QualifiedName lastSelectedValue_;
 }