+++ NEWS.md
@@ -6,9 +6,13 @@
 * Improvements to R code completion:
     - Completions provided automatically without an explicit gesture
     - All objects visible from the current scope are now included in completions
-    - Completions available in more contexts including S3 and S4 methods and dplyr pipelines
+    - Completions in more contexts including S3 and S4 methods and dplyr pipelines
     - Automatic insertion of closing parens when appropriate
+    - Inline help tooltip for signatures of completed functions
     - Completion for statements spanning multiple lines
+    - Specialized autocompletions provided for `library`, `data`, `vignette`, `?`
+    - For Shiny applications, autocompletions for `ui.R` + `server.R` pairs
+    - Completions for dimension names in `[`, `[[` calls
 * Improvements to C/C++ editing mode:
     - Code completion
     - F2 code navigation (go to definition)
@@ -45,6 +49,7 @@
 * Bind Cmd+Shift+K shortcut to Compile PDF and Preview HTML
 * When evaluating R strings ensure 'try' is called from base package
 * Add Clear Recent Projects menu item to toolbar project menu
+* Command to sync current working directory and Files pane to project directory
 * Update to MSYS SSH 1000.18
 * Update to pandoc 1.13.1